DEV: - implemented model_lgbm_63 - fully tested
note 1 : this is a implementation of eda 67 that run with 063 (see history), this version is standardize and passing tests
note 2 : there were a lot of stuff to fix in test since i was using really small size for data training. Now (as a STANDARD) i will be using first 100k rows of market_train_df as market_train_head_df. TODO: extend the newly implemented rolling test to lgbm tepmlate model